Inline support.js now that it contains basically no code
I went with preamble.js to include this as that is where MINIMAL_RUNTIME
includes `runtime_debug.js` from.
