+++ src/preamble.js
@@ -1141,4 +1141,20 @@ function createWasm() {
 var tempDouble;
 var tempI64;
 
+var STACK_ALIGN = {{{ STACK_ALIGN }}};
+var POINTER_SIZE = {{{ MEMORY64 ? 8 : 4 }}};
+
+{{{ getNativeTypeSize }}}
+
+#include "runtime_debug.js"
+
+#if RETAIN_COMPILER_SETTINGS
+var compilerSettings = {{{ JSON.stringify(makeRetainedCompilerSettings()) }}} ;
+
+function getCompilerSetting(name) {
+  if (!(name in compilerSettings)) return 'invalid compiler setting: ' + name;
+  return compilerSettings[name];
+}
+#endif // RETAIN_COMPILER_SETTINGS
+
 // === Body ===
