+++ test/other/test_unoptimized_code_size.js.size
@@ -1,1 +1,1 @@
-68195
+68143
