+++ test/other/test_unoptimized_code_size_strict.js.size
@@ -1,1 +1,1 @@
-67061
+67009
